# convert body to *char and adjust the job's allocation