This fix is trivial and the only place needed adjustment was
initialization of the pointer in allocate_job.

Job is still allocated and freed as one block. Performance has
improved slightly, not sure if this just a statistical variance.
We lack proper benchmarking so I cannot say for sure.

Before:

```
ctbench_heap_insert	 3000000	       545 ns/op
ctbench_heap_remove	 1000000	      1003 ns/op
ctbench_make_job	 5000000	       272 ns/op
ctbench_put_delete_1k	    3000	    351086 ns/op	   3.06 MB/s
ctbench_put_delete_64k	    2000	    617025 ns/op	 128.07 MB/s
ctbench_put_delete_8	    5000	    342281 ns/op	   0.04 MB/s
ctbench_put_delete_8k	    3000	    413185 ns/op          24.00 MB/s
```
After:
```
ctbench_heap_insert	 3000000	       546 ns/op
ctbench_heap_remove	 2000000	      1041 ns/op
ctbench_make_job	 5000000	       254 ns/op
ctbench_put_delete_1k	    5000	    320182 ns/op	   4.83 MB/s
ctbench_put_delete_64k	    2000	    609554 ns/op	 128.26 MB/s
ctbench_put_delete_8	    5000	    312411 ns/op	   0.04 MB/s
ctbench_put_delete_8k	    5000	    371577 ns/op	  37.72 MB/s
```
Fixes #487.