@ysmolsky 2019-07-27
This brings us 1 error closer to `-pedantic` flag.