@kr 2019-07-27
> Job is still allocated and freed as one block. Performance has
> improved slightly, not sure if this just a statistical variance.

I believe it should be possible to run https://godoc.org/golang.org/x/perf/cmd/benchstat on the output of our benchmarks. With the output of several benchmark runs, that tool can show the statistical significance of the results.