@ysmolsky 2019-07-28
> I believe it should be possible to run https://godoc.org/golang.org/x/perf/cmd/benchstat on the output of our benchmarks. With the output of several benchmark runs, that tool can show the statistical significance of the results.

Thanks for reviewing. I know that it is possible, just not convenient right now. `_ctcheck` outputs tests results as well. And _ctcheck could run benchmark suite N times too.