@kr 2019-07-28
Oh yeah good point.